hacking for january demo
split up present processing so that:

i) feed in a small known CSV file
ii) code notices that this is a known demo CSV file, and simulates discover of a similar larger dataset
 ii.1) this should appear as some user interface stage that looks like "Discovering similar data"
 ii.2) then rest of blindml run happens with that "similar" (i.e. full CSV file)
iii) rest of demo is as before

To start with, break out run_wit into separate steps in a notebook, I guess?

Do I need to change to a different dataset? I think its ok to use the existing perovskite one (but cut up as above)